+++ R/importancies.R
@@ -5,8 +5,6 @@
 ### Useful to analyze feature importance
 
 
-# labcolor <- paste0("<i style='color:",labcolor,"'>",names,"</i>")
-
 #' Importance barplot
 #'
 #' `plotImp()` displays the barplot of a numeric vector, which is assumed to contain the
@@ -88,11 +86,14 @@ plotImp <- function(x,y=NULL, relative=TRUE, absolute=TRUE, nfeat=NULL,
   x_abs <- x_abs[xord]
 
   if(!is.null(nfeat)) {
+    if(nfeat>length(x)) stop("nfeat cannot be larger than x")
     x <- x[1:nfeat]
     x_abs <- x_abs[1:nfeat]
   }
 
+
   xord <- order(x_abs,decreasing = FALSE)
+
 
   if(relative) {
     x_abs <- x_abs/x_den
@@ -109,12 +110,21 @@ plotImp <- function(x,y=NULL, relative=TRUE, absolute=TRUE, nfeat=NULL,
     y <- y[xord]
   }
 
+  if(length(color)==length(names))  {
+    names(color) <- names
+    color <- color[names(x_abs)]
+    color <- color[xord]
+  }
+
   if(absolute) {
-    if(is.null(xlim)) xlim <- c(0,max(c(x,y))+0.1*max(c(x,y)))
+    if(is.null(xlim)) xlim <- c(0,max(c(x_abs[xord],y))+0.1*max(c(x_abs[xord],y)))
     p <- graphics::barplot(x_abs[xord],las=2,horiz=TRUE, main=main,xlim=xlim, border=color, col=color,...)
     names <- names(x_abs[xord])
   } else {
-    if(is.null(xlim)) xlim <- c(min(c(y,x))-0.1*abs(min(c(y,x))),max(c(x,y))+0.1*max(c(x,y)))
+    if(is.null(xlim)) xlim <- c(min(c(y,x[xord]))-0.1*abs(min(c(y,x[xord]))),max(c(x[xord],y))+0.1*max(c(x[xord],y)))
+    signs <- c(sign(xlim[1]),sign(xlim[2]))
+    if(all(signs==c(-1,-1))) xlim[2] <- 0
+    if(all(signs==c(1,1))) xlim[1] <- 0
     p <- graphics::barplot(x[xord],las=2,horiz=TRUE, main=main,xlim=xlim, border=color, col=color,...)
     names <- names(x[xord])
   }
@@ -136,4 +146,60 @@ plotImp <- function(x,y=NULL, relative=TRUE, absolute=TRUE, nfeat=NULL,
   toreturn <- list(first_features=names(x), cumsum=sum(x_abs),barplot = p)
   graphics::par(mar = original_mar)
   return(toreturn)
+}
+# labcolor <- paste0("<i style='color:",labcolor,"'>",names,"</i>")
+
+
+
+#' Aggregate importances
+#'
+#' `aggregate_imp()` sums the importances present in a matrix or data.frame according
+#' to some user-specified grouping criterion.
+#'
+#' @param X Matrix or data.frame containing the importances (in rows or in columns).
+#' @param samples Samples are in rows or in columns? (Defaults: "rows").
+#' @param lev (optional) The grouping elements. `lev` should be as long as the dimension (cols or rows)
+#' that one wants to aggregate. If this parameter is absent, the colnames (if samples="rows") or rownames will
+#' be used to that effect. In that case, it is expected that the col/rownames follow
+#' this pattern: "V_Y", and the variables with the same "V" will be summed. (Check the colnames of
+#' a typical output of `dummy_data()` for more info).
+#' @return X, a matrix or data.frame containing the aggregated importances.
+#'
+#' @importFrom methods hasArg
+#' @importFrom dplyr %>% group_by id summarise_all
+#' @export
+#'
+#' @examples
+#' importances <- matrix(rnorm(90),nrow=3,ncol=30)
+#' rownames(importances) <- c("sample1","sample2","sample3")
+#' colnames(importances) <- paste0("Feat",
+#' rep(1:5,times=2*(1:5)), "_", unlist(lapply(2*(1:5),function(x)LETTERS[1:x])))
+#'
+#' ## The grouping criterion is:
+#' groups <- paste0("Feat",1:5)
+#' aggregate_imp(X=importances,samples="rows",lev=groups)
+
+#' ## We can also use the colnames:
+#' colnames(importances)
+#' aggregate_imp(X=importances,samples="rows")
+
+aggregate_imp <- function(X,lev=NULL,samples="rows") {
+  if(samples=="rows") X <- t(X)
+  if(!hasArg(lev)) {
+    lev <- sub('_.*', '', rownames(X))
+  } else {
+    ids <- sapply(lapply(paste0(lev,"_"),function(x)grep(x,rownames(X))),length)
+    lev <- rep(lev,times= ids )
+  }
+  levs <- NULL
+  group_id <- data.frame(X,levs=lev) %>% group_by(levs) %>% summarise_all(sum)
+  group_id <- as.data.frame(group_id)
+  colsnames <- group_id$levs
+  group_id <- t(group_id[,-1])
+  colnames(group_id) <- colsnames
+  if(samples=="rows") {
+      return(group_id[,unique(lev)])
+  } else {
+    return(t(group_id[,unique(lev)]))
+  }
 }