+++ tests/testthat/test-importancies.R
@@ -21,4 +21,27 @@ test_that("plotImp works", {
 
   plotimp <- plotImp(x,absolute=TRUE,relative=FALSE,names=1:30)
   expect_equal(plotimp$first_features, sub("Feat","",names(sort(abs(x),decreasing = TRUE))))
+
+  plotimp <- plotImp(x,absolute=TRUE,relative=FALSE,names=1:30)
+  expect_error( plotImp(x,nfeat=50), "nfeat cannot be larger than x")
+})
+
+
+test_that("aggregate_imp works", {
+  vect_grep <- Vectorize(grepl)
+  importances <- matrix(rnorm(120),nrow=4,ncol=30)
+  rownames(importances) <- paste0("sample",1:4)
+  colnames(importances) <- paste0("Feat",rep(1:5,times=2*(1:5)), "_", unlist(lapply(2*(1:5),function(x)LETTERS[1:x])))
+
+  aggrimp <- aggregate_imp(X=importances,samples="rows")
+  expect_equal(aggrimp[,"Feat3"],rowSums(importances[,vect_grep(sub('_.*', '',colnames(importances)),"Feat3")]))
+  expect_equal(aggrimp[,"Feat5"],rowSums(importances[,vect_grep(sub('_.*', '',colnames(importances)),"Feat5")]))
+  expect_equal(t(aggrimp), aggregate_imp(X=t(importances),samples="cols"))
+
+  groups <- paste0("Feat",1:5)
+  expect_equal(aggrimp, aggregate_imp(X=importances,lev=groups))
+  groups <- paste0("Feat",2:4)
+  expect_error(aggregate_imp(X=importances,lev=groups),"arguments imply differing number of rows")
+  expect_error(aggregate_imp(X=importances,lev=31:35),"arguments imply differing number of rows")
+
 })